using Long instead of int in counter

+++ src/main/java/ca/queensu/cs/server/Event.java
@@ -20,7 +20,7 @@ public class Event implements Comparable<Event> {
 	private String eventTarget;
 	private String vectorTime;
 	private String regionName;
-	private Integer counter;
+	private Long counter;
 	
 	
 	/*
